Hoist Variable row out of index and use JsonInput instead of JSONInput
